+++ core/src/main/java/io/questdb/cairo/arr/DirectArray.java
@@ -75,7 +75,7 @@ public class DirectArray extends ArrayView implements ArraySink, Mutable {
         int byteSize = flatLength << ColumnType.pow2SizeOf(ColumnType.decodeArrayElementType(type));
         ensureCapacity(byteSize);
         this.flatViewLength = flatLength;
-        int elemType = ColumnType.decodeArrayElementType(type);
+        short elemType = ColumnType.decodeArrayElementType(type);
         borrowedFlatView().of(ptr, elemType, flatLength);
         int stride = 1;
         for (int i = shape.size() - 1; i >= 0; i--) {
